# docs/swagger.yaml
basePath: /
definitions:
...
  model.TrackResponse:
    properties:
      album:
        type: string
      artist:
        type: string
      createdAt:
        type: string
      duration:
        type: integer
      genre:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  model.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@music-service.com
    name: API Support
    url: http://example.com/support
  description: API для сервиса стриминга и селфхостинга музыки
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Music Streaming Service API
  version: "1.0"
paths:
...
  /api/tracks:
    get:
      description: Возвращает список всех треков в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TrackResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получить все треки
      tags:
      - Tracks
    post:
      consumes:
      - multipart/form-data
      description: Загружает аудиофайл и создает запись о треке
      parameters:
      - description: Аудиофайл
        in: formData
        name: file
        required: true
        type: file
      - description: Название трека
        in: formData
        name: title
        required: true
        type: string
      - description: Исполнитель
        in: formData
        name: artist
        required: true
        type: string
      - description: Альбом
        in: formData
        name: album
        type: string
      - description: Жанр
        in: formData
        name: genre
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Загрузить новый трек
      tags:
      - Tracks
  /api/tracks/{id}:
    delete:
      description: Удаляет трек по ID (только для владельца)
      parameters:
      - description: ID трека
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Удалить трек
      tags:
      - Tracks
    get:
      description: Возвращает информацию о конкретном треке
      parameters:
      - description: ID трека
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получить трек по ID
      tags:
      - Tracks
  /api/tracks/search:
    get:
      description: Поиск треков по названию, исполнителю, альбому или жанру
      parameters:
      - description: Поисковый запрос
        in: query
        name: q
        type: string
      - description: Исполнитель
        in: query
        name: artist
        type: string
      - description: Альбом
        in: query
        name: album
        type: string
      - description: Жанр
        in: query
        name: genre
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TrackResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Поиск треков
      tags:
      - Tracks
  /api/tracks/stream/{id}:
    get:
      description: Возвращает аудиопоток для проигрывания трека
      parameters:
      - description: ID трека
        in: path
        name: id
        required: true
        type: integer
      produces:
      - audio/mpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Воспроизвести трек
      tags:
      - Tracks
  /api/user/profile:
    get:
      description: Возвращает информацию о текущем авторизованном пользователе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получить профиль пользователя
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Обновляет информацию о текущем авторизованном пользователе
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Обновить профиль пользователя
      tags:
      - User
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"